+++ lib/calendar_assistant/config.rb
@@ -33,13 +33,14 @@ class CalendarAssistant
       module Options
         COMMITMENTS = "commitments" # bool
         JOIN = "join"               # bool
-        ATTENDEES = "attendees"     # array of calendar ids (comma-delimited)
+        CALENDARS = "calendars"     # array of calendar ids (comma-delimited)
         LOCAL_STORE = "local-store" # filename
         DEBUG = "debug"             # bool
         FORMATTING = "formatting"   # Rainbow
         MUST_BE = "must-be"         # array of event predicates (comma-delimited)
         MUST_NOT_BE = "must-not-be" # array of event predicates (comma-delimited)
         CONTEXT = "context"         # symbol referring to command context
+        FORCE = "force"             # bool
       end
     end
 
@@ -50,7 +51,7 @@ class CalendarAssistant
       Keys::Settings::MEETING_LENGTH => "30m",            # ChronicDuration
       Keys::Settings::START_OF_DAY => "9am",              # BusinessTime
       Keys::Settings::END_OF_DAY => "6pm",                # BusinessTime
-      Keys::Options::ATTENDEES => [DEFAULT_CALENDAR_ID],  # array of calendar ids
+      Keys::Options::CALENDARS => [DEFAULT_CALENDAR_ID],  # array of calendar ids
       Keys::Options::FORMATTING => true,                  # Rainbow
     }
 
@@ -145,10 +146,10 @@ class CalendarAssistant
     end
 
     #
-    #  helper method for Keys::Options::ATTENDEES
+    #  helper method for Keys::Options::CALENDARS
     #
-    def attendees
-      split_if_array(Keys::Options::ATTENDEES)
+    def calendar_ids
+      split_if_array(Keys::Options::CALENDARS)
     end
 
     def must_be