+++ spec/calendar_assistant/event_repository_factory_spec.rb
@@ -3,9 +3,33 @@ describe CalendarAssistant::EventRepositoryFactory do
     let(:service) { instance_double "Service" }
     let(:calendar_id) { instance_double "calendar_id" }
 
-    it "creates an EventRepository" do
-      expect(CalendarAssistant::EventRepository).to receive(:new).with(service, calendar_id, anything)
-      CalendarAssistant::EventRepositoryFactory.new_event_repository(service, calendar_id)
+    context "when no type is set" do
+      it "creates an EventRepository" do
+        expect(CalendarAssistant::EventRepository).to receive(:new).with(service, calendar_id, anything)
+        described_class.new_event_repository(service, calendar_id)
+      end
+    end
+
+    context "when a type is set that is nil or weird" do
+      it "creates an EventRepository" do
+        expect(CalendarAssistant::EventRepository).to receive(:new).with(service, calendar_id, anything).twice
+        described_class.new_event_repository(service, calendar_id, type: :something_strange)
+        described_class.new_event_repository(service, calendar_id, type: nil)
+      end
+    end
+
+    context "when the type is lint" do
+      it "creates a LintEventRepository" do
+        expect(CalendarAssistant::LintEventRepository).to receive(:new).with(service, calendar_id, anything)
+        described_class.new_event_repository(service, calendar_id, type: :lint)
+      end
+    end
+
+    context "when the type is location" do
+      it "creates a LocationEventRepository" do
+        expect(CalendarAssistant::LocationEventRepository).to receive(:new).with(service, calendar_id, anything)
+        described_class.new_event_repository(service, calendar_id, type: :location)
+      end
     end
   end
 end