+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 # Changelog for Calendar Assistant
 
+## unreleased:
+
+- Support multiple calendars `--calendars` to manage location events on multiple calendars [#63]
+
 ## v0.8.0 / 2019-01-28
 
 Features:
-
 - Most commands now allow filtering by property. See README for details on `--must-be` and `--must-not-be`. [#72]
 - Allow location event visibility to be set. [#17]
 - Allow command-specific settings. [#17]